Rename hassio coordinator module
