+++ homeassistant/components/hassio/coordinator.py
